feat: store url 변경 및 styledComponent 분리
